Refactor resistance module into modular components with improved naming
