+++ crates/cfd-2d/src/pressure_velocity/solver.rs
@@ -76,7 +76,7 @@ impl<T: RealField + Copy + FromPrimitive + Copy + LowerExp> PressureVelocitySolv
     ) -> cfd_core::error::Result<T> {
         // Step 1: Solve momentum equations for predicted velocity
         // Note: This needs proper integration with SimulationFields
-        // For now, creating a temporary fields structure
+        // For now, creating a fields structure buffer
         let mut temp_fields = SimulationFields::new(self.grid.nx, self.grid.ny);
         // Copy current state to fields
         for i in 0..self.grid.nx {
